Observe TaskRoleName when printing tasks
print-tasks and print-stacks will resolve references (e.g. CopyValue statements) reaching out to AWS CloudFormation. We like to split the roles used for that between read-only tasks like printing and doing perform-tasks which requires a lot of privilege. This change will have print-tasks and print-stacks observe the TaskRoleName parameter on the task